Brought in changes to wording for tests, and fixed any testing bugs

+++ tests/rt.sh
@@ -141,7 +141,6 @@ verify_testing() {
           else
             if [[ $PRETEST == false ]]; then
               cat ${LOG_DIR}/rt_${TEST_ID}.log >> ${REGRESSIONTEST_LOG}
-              echo "Test ${TEST_ID} PASSED" >> ${REGRESSIONTEST_LOG}
             fi
           fi
         fi
@@ -687,7 +686,9 @@ NEW_BASELINE=${STMP}/${USER}/FV3_RT/REGRESSION_TEST
 # Overwrite default RUNDIR_ROOT if environment variable RUNDIR_ROOT is set
 RUNDIR_ROOT=${RUNDIR_ROOT:-${PTMP}/${USER}/FV3_RT}/rt_$$
 mkdir -p ${RUNDIR_ROOT}
-rm ${PATHRT}/run_dir
+if [[ -e ${PATHRT}/run_dir ]]; then
+  rm ${PATHRT}/run_dir
+fi
 echo "Linking ${RUNDIR_ROOT} to ${PATHRT}/run_dir"
 ln -s ${RUNDIR_ROOT} ${PATHRT}/run_dir
 echo "Run regression test in: ${RUNDIR_ROOT}"
